requirements.txt
Django>=4.1,<5.0
djangorestframework>=3.13.1,<4.0
drf-yasg>=1.21.3,<2.0

coverage
